requirements/base.txt
aliyun-python-sdk-core==2.13.36
celery==5.2.7
diskcache==5.6.1
Django==4.2.1
django-cors-headers==3.14.0
django-db-connection-pool==1.2.2
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
itsdangerous==2.1.2
loguru==0.7.0
minio==7.1.14
mysqlclient==2.1.1
orjson==3.8.12
pillow==9.5.0
redis==4.5.5
SQLAlchemy==2.0.15
